Tried adding a typescript file. Turns out you need a whole lot more than just some basic index.js and index.html files to get that properly running.
But! I got it. I added some basic code that creates a set of languages strings and translates it.

The basic approach is to add an attribute to the HTML elements you want to be translated. Basically, go through all of these and find their translation in the list you provided, and switch them out. Some other fancy stuff with creating a select element with language options based off the available options in the array of strings.

If I were to do something differently, I would probably 1) create separate .string files or .json files to store the data and handle it a bit more eloquently. I would probably add some sort of unique attribute which corresponds to a key of what that copy is. Since this is going to be a static content website, it would be nice to just add identifiers to the different parts of my website, and then based on the selected language, populate the content based off a unique key like "intro-part-1", "about-me-part-2", etc.

I'm proud that I was able to accomplish this strictly using JavaScript and keeping the project as lean as possible for the time being, but accomplished some pretty powerful functionality just with some JavaScript functions.

Reference: https://benfrain.com/creating-a-language-switcher-in-javascript